Changes to UserController, added UserMapper and UserService tests
